# Change files to be deleted in batches instead of one-by-one