Deleting files in object storage one by one is too slow during vacuuming or domain media clean-up. There are batch APIs for S3, and updating rows in a batch instead of one-by-one should be faster, too.

Comparison of clearing out a batch of 100 media attachments:

|Before|After|
|--|--|
|21s|1.2s|

___

Fixes MAS-49

Fix #23073, close #23097