@metadaddy 2023-02-11
Hi @Gargron - I noticed this failed some lint tests. Shall I fix them in https://github.com/mastodon/mastodon/pull/23097?