@metadaddy 2023-02-11
Ah - I see you've refactored my PR somewhat. I'll leave it in your capable hands.