@smiba 2024-01-10
@ClearlyClaire Any reason this wasn't backported to v4.1.x? I guess you have to draw the line somewhere, but this change seems to work just fine on v4.1.11. We currently have it in use and has made the vacuum_scheduler roughly 400 times faster.